% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ACSN_Enrichment.R
\docType{data}
\name{ACSN_maps}
\alias{ACSN_maps}
\title{Atlas of Cancer Signalling Networks}
\format{A list of dataframes
\describe{
 \item{Apoptosis}{Map of apoptosis}
 \item{CellCycle}{Map of the cell cycle}
 \item{DNA_repair}{Map of DNA repair}
}}
\source{
\url{https://acsn.curie.fr/downloads.html}
}
\usage{
ACSN_maps
}
\description{
A dataset containing the six maps of ACSN: apoptosis, cell cycle, DNA reparation, EMT motility, survival, and the master map
}
\keyword{datasets}

