\name{ACSNEnrichment-package}
\alias{ACSNEnrichment-package}
\alias{ACSNEnrichment}
\docType{package}
\title{ACSNEnrichment}
\packageTitle{ACSNEnrichment}
}
\description{
\packageDescription{ACSNEnrichment}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ACSNEnrichment}
\packageIndices{ACSNEnrichment}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ACSNEnrichment}

Maintainer: \packageMaintainer{ACSNEnrichment}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
