% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ACSN_Enrichment.R
\name{enrichment}
\alias{enrichment}
\title{Gene enrichment analysis from ACSN maps or gmt files}
\usage{
enrichment(Genes = NULL, maps = list(Apoptosis = ACSN_apop_formatted,
  CellCycle = ACSN_cellcyc_formatted, DNA_repair = ACSN_dnarep_formatted,
  EMT_motility = ACSN_emtmotil_formatted, ACSN_master = ACSN_master_formatted,
  Survival = ACSN_survival_formatted), correction_multitest = FALSE,
  statistical_test = "fisher", min_module_size = 5, universe = "HUGO",
  threshold = 1)
}
\arguments{
\item{Genes}{Character vector of genes that should be tested for enrichment}

\item{maps}{list of maps generated by format_from_gmt. Names of element of list will be used to track modules. Default: tests on all acsn maps.}

\item{correction_multitest}{either FALSE, "bonferroni", "holm", "hochberg", "hommel", "BH", "fdr" (identical to BH), or "BY"}

\item{min_module_size}{will remove from the analysis all modules which are (strictly) smaller than threshold}

\item{universe}{Universe on which the statistical analysis should be performed. Can be either "HUGO","ACSN", or a character vector of genes.}

\item{threshold}{: maximal p-value (corrected if correction is enabled) that will be displayed}

\item{statistical}{test one of "fisher", "hypergeom"}
}
\description{
Compute and represent gene set enrichment from your data based on pre-saved maps from ACSN or user imported maps. The gene set enrichment can be run with hypergeometric test or Fisher exact test,
and can use multiple corrections. Visualization of data can be done either by barplots or heatmaps.
}
\author{
Eric Bonnet <eric.bonnet at curie.fr>, Paul Deveau <paul.deveau at curie.fr!>
Gene set enrichment analysis
}

