% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ACSN_Enrichment.R
\docType{data}
\name{enrichment_test}
\alias{enrichment_test}
\title{Result from enrichment test of "genes_test" on the ACSN maps}
\format{data.frame
\describe{
 \item{module}{Name of module}
 \item{genes_in_module}{Genes from genes_test in module}
 \item{p.value}{Uncorrected p-value}
 \item{p.value.corrected}{p-value corrected for multiple testing by Bonferroni correction}
 }}
\usage{
enrichment_test
}
\description{
Parameters: bonferroni correction, min module size = 5
}
\keyword{datasets}

