% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ACSN_Enrichment.R
\name{multisample_enrichment}
\alias{multisample_enrichment}
\title{Enrichment analysis function}
\usage{
multisample_enrichment(Genes_by_sample = NULL,
  maps = ACSNEnrichment::ACSN_maps, correction_multitest = "BH",
  statistical_test = "fisher", min_module_size = 5,
  universe = "map_defined", individual_threshold = 1,
  cohort_threshold = 1)
}
\arguments{
\item{Genes_by_sample}{List of character vectors. Each list element name should be a sample name,
and each character vector the set of genes to test for the sample.}

\item{maps}{list of maps generated by format_from_gmt. Default: tests on all acsn maps}

\item{correction_multitest}{either FALSE, "bonferroni", "holm", "hochberg", "hommel", "BH", "fdr" (identical to BH), or "BY"}

\item{statistical_test}{one of "fisher", "hypergeom"}

\item{min_module_size}{will remove from the analysis all modules which are (strictly) smaller than threshold}

\item{universe}{Universe on which the statistical analysis should be performed. Can be either "HUGO","ACSN"
(identical to "map_defined"), or a character vector of genes.}

\item{individual_threshold}{maximal p-value (corrected if correction is enabled) that will be displayed}

\item{cohort_threshold}{if TRUE modules will be kept in all samples if at least one sample has p-value lower than threshold, otherwise the threshold is applied for each sample independently.}
}
\description{
Allows gene set enrichment analysis
}

